Added Mock API for cross merchant recomms

+++ recommendation-api/src/main/java/com/simpl/sheldon/recommendationapi/models/commons/LineItem.java
@@ -16,7 +16,7 @@ public class LineItem {
     private String description;
 
     @JsonProperty("price")
-    private int price;
+    private long price;
 
     @JsonProperty("unit_price")
     private int unitPrice;
